Migrate Single Stock Data From CatalogInventory to MSI SourceItems

+++ app/code/Magento/InventoryApi/Test/_files/products_rollback.php
@@ -3,6 +3,10 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
+declare(strict_types=1);
+
+use Magento\Framework\Api\SearchCriteriaBuilder;
+use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Catalog\Api\ProductRepositoryInterface;
 use Magento\CatalogInventory\Api\StockStatusCriteriaInterfaceFactory;
 use Magento\CatalogInventory\Api\StockStatusRepositoryInterface;
@@ -23,17 +27,27 @@ $currentArea = $registry->registry('isSecureArea');
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', true);
 
-for ($i = 1; $i <= 3; $i++) {
-    $product = $productRepository->get('SKU-' . $i);
-    /** @var \Magento\CatalogInventory\Api\StockStatusCriteriaInterfaceFactory $stockStatusCriteriaFactory **/
-    $criteria = $stockStatusCriteriaFactory->create();
-    $criteria->setProductsFilter($product->getId());
+$searchCriteriaBuilder = Bootstrap::getObjectManager()->get(SearchCriteriaBuilder::class);
+$searchCriteria = $searchCriteriaBuilder->addFilter(
+    ProductInterface::SKU,
+    ['SKU-1', 'SKU-2', 'SKU-3'],
+    'in'
+)->create();
+$products = $productRepository->getList($searchCriteria)->getItems();
 
-    $result = $stockStatusRepository->getList($criteria);
-    $stockStatus = current($result->getItems());
-    $stockStatusRepository->delete($stockStatus);
-
-    $productRepository->deleteById('SKU-' . $i);
+/**
+ * Tests which are wrapped with MySQL transaction clear all data by transaction rollback.
+ * In that case there is "if" which checks that SKU1, SKU2 and SKU3 still exists in database.
+ */
+if (!empty($products)) {
+    foreach ($products as $product) {
+        $criteria = $stockStatusCriteriaFactory->create();
+        $criteria->setProductsFilter($product->getId());
+        $result = $stockStatusRepository->getList($criteria);
+        $stockStatus = current($result->getItems());
+        $stockStatusRepository->delete($stockStatus);
+        $productRepository->delete($product);
+    }
 }
 
 $registry->unregister('isSecureArea');
